leadmanager/requirements.txt
aiocontextvars==0.2.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
contextvars==2.4
cryptography==3.1.1
Django==2.2.10
django-rest-knox==4.1.0
djangorestframework==3.11.1
dynaconf==3.1.1
hvac==0.10.5
idna==2.10
immutables==0.14
Jinja2==2.11.2
knox==0.1.0
loguru==0.5.3
MarkupSafe==1.1.1
psycopg2==2.8.2
psycopg2-binary==2.8.6
pycparser==2.20
pytz==2018.9
requests==2.24.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.3.1
Unipath==1.1
urllib3==1.25.10
